RecipeGPT-Backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
moviepy==1.0.3
imageio==2.37.0
imageio-ffmpeg==0.4.8
requests==2.32.3
openai-whisper==20231117
torch==2.1.2
yt-dlp==2025.3.31